+++ lib/keyfile/keyfile.h
@@ -5,13 +5,14 @@
 #include <stdio.h>
 
 /**
- * keyfile_read(filename, machinenum, keys, force):
+ * keyfile_read(filename, machinenum, keys, force, devtty):
  * Read keys from a tarsnap key file; and return the machine # via the
  * provided pointer.  Ignore any keys not specified in the ${keys} mask.
  * If ${force} is 1, do not check whether decryption will exceed
- * the estimated available memory or time.
+ * the estimated available memory or time.  If ${devtty} is non-zero, read a
+ * password from /dev/tty if possible; if not, read from stdin.
  */
-int keyfile_read(const char *, uint64_t *, int, int);
+int keyfile_read(const char *, uint64_t *, int, int, int);
 
 /**
  * keyfile_write(filename, machinenum, keys, passphrase, maxmem, cputime):